pyproject.toml
[tool.poetry]
name = "online-training"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = """(/(|/.git)/)"""


[tool.isort]
line_length = 119
